+++ src/main/java/command/CommandCollections.java
@@ -1,0 +1,5 @@
+package command;
+
+public enum CommandCollections {
+    LIST, TODO, DEADLINE, EVENT, DONE, DELETE, BYE, VIEW, SEARCH, HELP, RESET
+}
