+++ src/main/java/command/DoneCommand.java
@@ -1,0 +1,24 @@
+package command;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+import java.io.IOException;
+
+public class DoneCommand extends Command{
+    String[] command;
+
+    public DoneCommand(String[] command){
+        this.command = command;
+    }
+
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws IOException {
+        int taskNumber = Integer.parseInt(command[1]) - 1;
+        taskList.get(taskNumber).setDone(true);
+        ui.printDone(taskNumber, taskList);
+        ui.printTaskCount(taskList);
+        storage.saveTask(taskList.getTasks());
+    }
+
+}
