+++ src/main/java/command/HelpCommand.java
@@ -1,0 +1,15 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+import java.io.IOException;
+
+public class HelpCommand extends Command{
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, IOException {
+        ui.howToUse();
+    }
+}
