+++ src/main/java/command/InvalidCommand.java
@@ -1,0 +1,18 @@
+package command;
+
+import exception.DukeException;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+public class InvalidCommand extends Command{
+    private final String errorMsg;
+
+    public InvalidCommand(String errorMsg){
+        this.errorMsg = errorMsg;
+    }
+
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
+        throw new DukeException(errorMsg);
+    }
+}
