+++ src/main/java/command/ListCommand.java
@@ -1,0 +1,15 @@
+package command;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+public class ListCommand extends Command {
+
+    public ListCommand (){}
+
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        ui.printTaskList(taskList);
+        ui.printTaskListStr(taskList);
+    }
+}
