+++ src/main/java/command/ViewCommand.java
@@ -1,0 +1,24 @@
+package command;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+import java.time.LocalDate;
+
+/**
+ * The <code>viewCommand</code> class
+ */
+public class ViewCommand extends Command {
+    LocalDate date;
+
+    public ViewCommand(LocalDate date){
+        this.date = date;
+    }
+
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        TaskList taskByDate = taskList.getTaskByDate(date, taskList);
+        ui.printTaskList(taskByDate);
+        ui.printTaskOfDate(taskByDate, date);
+    }
+}
