--- src/main/java/duke/command/CommandCollections.java
@@ -1,5 +1,0 @@
-package duke.command;
-
-public enum CommandCollections {
-    LIST, TODO, DEADLINE, EVENT, DONE, DELETE, BYE, VIEW, SEARCH, HELP, RESET
-}
