--- src/main/java/duke/command/HelpCommand.java
@@ -1,15 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-import java.io.IOException;
-
-public class HelpCommand extends Command{
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, IOException {
-        ui.howToUse();
-    }
-}
