--- src/main/java/duke/command/ListCommand.java
@@ -1,15 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.tasklist.*;
-import duke.ui.Ui;
-
-public class ListCommand extends Command {
-
-    public ListCommand (){}
-
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        ui.printTaskList(taskList);
-        ui.printTaskListStr(taskList);
-    }
-}
