--- src/main/java/duke/command/ResetCommand.java
@@ -1,19 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-import java.io.IOException;
-
-public class ResetCommand extends Command{
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, IOException {
-        taskList.reset();
-        ui.reset();
-        ui.printTaskCount(taskList);
-        storage.saveTask(taskList.getTasks());
-    }
-}
