+++ src/main/java/exception/DukeException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class DukeException extends IllegalArgumentException{
+    public DukeException(String s) {
+        super(s);
+    }
+}
